+++ packages/core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @veloxts/core
 
+## 0.6.49
+
+### Patch Changes
+
+- feat(create): add client import lint script for RSC templates
+
 ## 0.6.48
 
 ### Patch Changes